points -> long instead of int, retailer name points

+++ src/main/java/com/allengueco/receipt/processor/AbstractProcessor.java
@@ -3,5 +3,5 @@ package com.allengueco.receipt.processor;
 import com.allengueco.receipt.model.Receipt;
 
 public interface AbstractProcessor {
-    int process(Receipt receipt);
+    long process(Receipt receipt);
 }
